refactor: util 파일 위치 수정
